backend/pyproject.toml
[project]
name = "Kodomo-Wallet-Backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "alembic>=1.17.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.119.1",
    "httpx>=0.28.1",
    "injector>=0.22.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.44",
    "strawberry-graphql>=0.284.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.8.0",
]

[tool.ruff]
# Python 3.14をターゲット
target-version = "py314"

# 1行の最大文字数
line-length = 100

# 除外するディレクトリ
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
    ".pytest_cache",
    "migrations",
]

[tool.ruff.lint]
# 有効にするルールセット
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import順序)
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]

# 無視するルール
ignore = [
    "E501",  # line-too-long (line-lengthで制御)
    "B008",  # do-not-perform-function-calls-in-argument-defaults (FastAPIのDepends等で必要)
    "ARG001", # unused-function-argument (オーバーライドで必要な場合がある)
]

# ファイルごとのルール無視設定
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused-import (re-exportで使用)
"test_*.py" = ["ARG001", "SIM117"]  # テストファイルでは緩和

[tool.ruff.lint.isort]
# import順序の設定
known-first-party = ["app"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
# フォーマットスタイル
quote-style = "double"
indent-style = "space"
line-ending = "auto"
